continue to improve IA

+++ include/game/systems/Brain.hpp
@@ -37,14 +37,14 @@ namespace game
                 void updateMaps(void);
 
                 double _timer;
-                Vector<int> _goal{0, 0};
+                Vector<float> _goal{0, 0};
                 Vector<int> _sizeMap;
             private:
 
                 // * Environment
                 bool isADangerousZone(Vector<int> &pos);
-                bool isDangerous(Vector<int> &pos, Vector<int> dir = {0, 0});
-                bool isSolid(Vector<int> &pos, Vector<int> &dir);
+                bool isDangerous(Vector<int> &pos, Vector<float> dir = {0, 0});
+                bool isSolid(Vector<int> &pos, Vector<float> &dir);
                 bool isSolid(int x, int y);
                 bool isWall(int x, int y);
                 void updateDangerousMap(void);
@@ -52,7 +52,7 @@ namespace game
                 void computeRangeBomb(Vector<int> &bomb);
 
                 // * Strategy
-                void setNewGoal(Vector<int> &pos, Vector<int> &goal);
+                void setNewGoal(Vector<int> &pos, Vector<float> &goal);
                 bool needDropBomb(void);
                 game::Event getEventFromDirection();
 
@@ -60,11 +60,11 @@ namespace game
                 // * OFFENSE
                 short PathFinding(Vector<int> &pos, Vector<int> &goal);
                 void computeDirection(void);
-                void setNewGoalOffense(Vector<int> &pos, Vector<int> &goal);
+                void setNewGoalOffense(Vector<int> &pos, Vector<float> &goal);
 
                 // * Defense
                 short PathFinding(Vector<int> &pos);
-                void setNewGoalDefense(Vector<int> &pos, Vector<int> &goal);
+                void setNewGoalDefense(Vector<int> &pos, Vector<float> &goal);
                 void defensePathFinding(int x, int y, int d, int a, int b);
 
                 // * Attributs
@@ -80,16 +80,16 @@ namespace game
                 std::vector<std::vector<Tag>> _tagMap;
                 std::vector<std::vector<game::Event>> _directionMap;
                 std::vector<std::vector<int>> _distanceMap;
-                Vector<int> _direction{0, 0};
+                Vector<float> _direction{0, 0};
                 int _level;
                 gameEngine::component::Clock _clock;
-                const std::array<Vector<int>, 5> directions =
+                const std::array<Vector<float>, 5> directions =
                 {
-                    Vector<int>{-0, 0},
-                    Vector<int>{-1, 0},
-                    Vector<int>{1, 0},
-                    Vector<int>{0, 1},
-                    Vector<int>{0, -1}
+                    Vector<float>{-0, 0},
+                    Vector<float>{-1, 0},
+                    Vector<float>{1, 0},
+                    Vector<float>{0, 1},
+                    Vector<float>{0, -1}
                 };
 
                 const std::array<game::Event, 5> events =
